// frontend/src/types/accountInsight.ts
export interface AccountInsight {
  id: string;
  companyId: string;
  igId: string;
  date: Date;
  followers: number;
  follows: number;
  posts: number;
  reachTotal: number;
  reachFollower: number;
  reachNonFollower: number;
  reachUnknown: number;
  impressions: number;
  profileVisits: number;
  webTaps: number;
  likesTotal: number;
  commentsTotal: number;
  savesTotal: number;
  sharesTotal: number;
  adReach: number;
  adLikes: number;
  adComments: number;
  adSaves: number;
  adShares: number;
  feedReach: number;
  feedLikes: number;
  feedComments: number;
  feedSaves: number;
  feedShares: number;
  storyReach: number;
  storyLikes: number;
  storyComments: number;
  storySaves: number;
  storyShares: number;
  reelReach: number;
  reelLikes: number;
  reelComments: number;
  reelSaves: number;
  reelShares: number;
  carouselReach: number;
  carouselLikes: number;
  carouselComments: number;
  carouselSaves: number;
  carouselShares: number;
  createdAt: Date;
  updatedAt: Date;
} 